Restructure components and remove old modal window
