--- components/RegionDialog/index.tsx
@@ -1,2 +1,0 @@
-export * from "./RegionDialog";
-export { default } from "./RegionDialog";
