+++ components/regionDropdowns/index.tsx
@@ -1,0 +1,2 @@
+export * from "./ChannelDropdown";
+export * from "./LocaleDropdown";
